refactor(ad): split ad_utils into _ad_primitives, breaking CTM SCC
Item 1 of #351. ``ad_utils`` participated in the 10-module CTM SCC
because:

  - ``ad_utils`` imports ``_ctm_tensor`` and ``_split_ctm_tensor``
    shims at module level.
  - ``_ctm_projector`` imported ``_fix_svd_signs``, ``truncated_svd_ad``,
    ``regularized_svd``, ``regularized_eigh`` from ``ad_utils`` (via
    deferred imports inside functions, but still tracked by the static
    AST analyzer).

These two edges closed a cycle ``ad_utils → _ctm_tensor → _ctm_projector
→ ad_utils``.  Closing it required moving the protocol-level (CTM-free)
AD pieces out so ``_ctm_projector`` can depend on them without going
through ``ad_utils``.

Changes:

  - New module ``tenax.algorithms._ad_primitives`` containing
    ``CTMRGGradientError``, ``_fix_svd_signs``, ``truncated_svd_ad``
    (+ helpers), ``truncated_svd_ad_vh_only``, ``regularized_svd``,
    ``regularized_eigh``, ``truncated_svd_symmetric_ad``, and
    ``_svd_sector_backward``.  No CTM imports.
  - ``ad_utils`` re-exports those names so external callers and tests
    keep working unchanged.
  - ``_ctm_projector`` imports the four primitives directly from
    ``_ad_primitives``.
  - Test monkey-patch sites updated to patch ``_ad_primitives`` (the
    actual module the projector reads from).
  - Architecture-import allowlist emptied — the previous 10-module
    legacy SCC is fully gone.

After this PR the static algorithm-package import graph has zero
SCCs of size > 1.
